Revert `new` as expression
